Don't initialize OpenGL until the last minute. This works around a situation observed with certain driver versions on Intel HD Graphics on Windows 7 where a driver thread eats CPU from when we initalize OpenGL until the first frame is rendered. Sadly this workaround means there is now an uncomfortably long delay after Preview is clicked for the very first time.
Also, abandon the business of making the screensaver window the owner
of the dialog window. Set the dialog window's icons directly and omit
WS_SYSMENU from the dialog's window styles. This gives exactly the
same effect, except that now the dialog has the correct icon in the
task switcher window.