Decrease the friction parameter for collisions with the walls.
At low density (when collisions between two objects are infrequent),
the objects tend to reach a state in which they bounce between one or
two pairs of opposite walls and never collide with each other. This
isn't very interesting to watch, so tweak the collision parameter to
make it less likely.